requirements.txt
cffi==1.17.0
cryptography==43.0.0
numpy==2.0.1
pillow==10.4.0
pycparser==2.22
tk==0.1.0
